Add the feature flag for scheduleManager
