+++ addons/schedule/common/src/main/java/org/commonjava/indy/schedule/conf/ScheduleDBConfig.java
@@ -88,5 +88,5 @@ public class ScheduleDBConfig implements IndyConfigInfo
     {
         return null;
     }
-    
+
 }
